requirements.txt
fastapi
uvicorn
databases[postgresql]
asyncpg
python-dotenv